+++ srai/loaders/osm_loaders/osm_tile_loader.py
@@ -15,14 +15,13 @@ import requests
 from srai._optional import import_optional_dependencies
 from srai.geodatatable import GeoDataTable
 from srai.loaders._base import VALID_AREA_INPUT, Loader
-from srai.regionalizers.slippy_map_regionalizer import SlippyMapRegionalizer
-
-from .osm_tile_data_collector import (
+from srai.loaders.osm_loaders.osm_tile_data_collector import (
     DataCollector,
     DataCollectorType,
-    InMemoryDataCollector,
+    SavingDataCollector,
     get_collector,
 )
+from srai.regionalizers.slippy_map_regionalizer import SlippyMapRegionalizer
 
 
 class OSMTileLoader(Loader):
@@ -46,7 +45,7 @@ class OSMTileLoader(Loader):
         resource_type: str = "png",
         auth_token: Optional[str] = None,
         data_collector: Optional[Union[str, DataCollector]] = None,
-        storage_path: Optional[Union[str, Path]] = None,
+        storage_path: Union[str, Path] = "files/osm_tiles",
     ) -> None:
         """
         Initialize TileLoader.
@@ -60,11 +59,12 @@ class OSMTileLoader(Loader):
             auth_token (str, optional): auth token. Added as access_token parameter
                 to request. Defaults to None.
             data_collector (Union[str, DataCollector], optional): DataCollector object or
-            enum defining default collector. If None uses InMemoryDataCollector. Defaults to None.
-            If `return` uses  InMemoryDataCollector
-            If `save` uses  SavingDataCollector
+                enum defining default collector. If None uses InMemoryDataCollector.
+                Defaults to None.
+                If `return` uses  InMemoryDataCollector.
+                If `save` uses  SavingDataCollector.
             storage_path (Union[str, Path], optional): path to save data,
-                used with SavingDataCollector. Defaults to None.
+                used with SavingDataCollector. Defaults to "files/osm_tiles".
 
         References:
             1. https://wiki.openstreetmap.org/wiki/Slippy_map_tilenames
@@ -79,7 +79,7 @@ class OSMTileLoader(Loader):
         self.data_collector = (
             self._get_collector(data_collector)
             if data_collector is not None
-            else InMemoryDataCollector()
+            else SavingDataCollector(save_path=self.save_path, file_extension=self.resource_type)
         )
         self.regionalizer = SlippyMapRegionalizer(zoom=self.zoom)